chore: better nav stack logic
